src/requirements.txt
tensorflow
fastapi
uvicorn
pillow
scikit-learn
python-multipart
psycopg2-binary
numpy
pandas
matplotlib
seaborn
dotenv
